+++ Minesweeper/Core/Cells/CellState.cs
@@ -1,0 +1,7 @@
+﻿namespace Minesweeper.Core
+{
+    public enum CellState
+    {
+        Opened, Closed
+    }
+}
