+++ Minesweeper/Core/Cells/CellType.cs
@@ -1,0 +1,7 @@
+﻿namespace Minesweeper.Core
+{
+    public enum CellType
+    {
+        Regular, Mine, Flagged, FlaggedMine
+    }
+}
