+++ Minesweeper/Core/GameMode.cs
@@ -1,0 +1,7 @@
+﻿namespace Minesweeper.Core
+{
+    public enum GameMode
+    {
+        Beginner, Intermediate, Expert
+    }
+}
